# Discussion: Make setBlocks update entities like setBlock - at huge performance cost !!!