@GooeyHub 2016-09-12
Hooray Jenkins reported success with all tests good!