@skaldarnar 2016-09-12
I suppose the life cycle events for blocks were removed/not added due to performance reasons - would be worth another benchmark to estimate the impact. I like the idea of an `OnChangedBlockEvent`, it's quite common to act on block changes, having a proper event can simplify working with it. 
Point (3) sounds like a reasonable change as it clearifies the purpose of `EntityAwareWorldProvider` imho.

I will test the changes as soon as I can find the time for it. :wink: