@Cervator 2016-09-13
How exactly does this relate to #2474 (empty chest - but even with a single GQ dungeon, no overlapping needed) or #2361 (same exact error message, but at least a couple different scenarios) ?

Although I noticed https://github.com/Terasology/StructureTemplates/commit/24e806062b900fae3f0950d6799e4863bc3d309e - does that on its own fix #2474 ? :-) If it is/was a StructureTemplates issue instead of engine that makes sense.