@flo 2016-09-13
About 3.) I found out that it is actually already keeping track of the block entities via component activation/deactivation. The temporary block entities are however separate. 

I really wonder if there is a nice minimalistic approach to get the performance we want.

Maybe we could simply skip the creation of the temporary block entities when a block gets placed. This would mean that of course there are no events that fake the conversion of a block entity from one type to another. It destroys a bit of the illusion that all blocks are entities. However that illusion isn't currently complete either, as you can't for example find a block via component when it has currently no temporary block entity.