@flo 2017-01-19
I would suggest we keep this PR for a while. I still want to fix it at some point.

In case someone else wants to do it in the meantime:

It might be enough to limit the entity creation to entities that stay active.