@flo 2017-05-08
Closing it now as the bug no longer seems reproducable.