@Cervator 2017-05-09
Huh, it went away on its own? Am thrilled to have those closed but puzzled on what happened :-)

Any risk it may take part in other odd issues like #2516? I wouldn't want us to walk away from a potential bug/fix from it just having sneakily hid its role in other problems.